requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
django-redis
celery==5.4.0
redis==5.2.1
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-celery-beat